[ios] Add dark mode color for ntp_background_color colorset
ntp_background_color is used as the background color for the fake
omnibox when pinned to the top of the NTP. This change adds a dark
mode color for it since it will automatically toggle when the device
settings changes. Otherwise, it is very difficult for the header logic
to update the correct coloring on its own since it doesn't have
internal awareness of whether it is pinned to the top.

Bug: 1432252
Change-Id: Id646dce0356826644cb31d5e82149bdb7ef428b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4510061
Reviewed-by: Scott Yoder <scottyoder@google.com>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141048}