Rename blink::Color::FloatAlpha() to Alpha()
As part of the CSS Color 4/5 efforts, blink::Colors now store floats
as color channels and alpha. FloatAlpha is thus a simple getter and
per the style guide should be called "Alpha()".

There was a Color::Alpha() that returned an integer version of
alpha in the range [0,255]. To be super clear, the name of that function
was changed to GetAlphaAsInteger() in https://chromium-review.googlesource.com/c/chromium/src/+/4494556.

GetAlphaAsInteger is being used in a lot of places where IsTransparent
would be more appropriate. This will be addressed in a future CL.

Bug: 1434423, 1399566
Change-Id: I2ca4da2dc41dc6484dbdf8a63cb05f2520bfd3de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4496486
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141071}