[fuchsia] Set abi_revision in resolved Components
Report the Fuchsia platform ABI revision targeted by resolved Cast
components as that with which the Cast resolver implementation was
built. This will be used in future to ensure that components outside
the compatibility window of the platform build on which they are run
can be flagged as potentially incompatible, or prevented from launching.

Several APIs used by Cast components are actually passed-through to the
web_instance.cm in which the component's web content is hosted, such as
the `fuchsia.logger.LogSink` to which its script output is routed. It is
assumed that the platform ABI targeted by the web_instance.cm will also
be verified, and reported, so that that need not be reflected in the
revision reported by the Cast Resolver.

Bug: fuchsia:119314
Change-Id: I4cea755b0ba61bd1a759d9b16f6345a71849aab0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4507148
Commit-Queue: Wez <wez@chromium.org>
Reviewed-by: David Dorwin <ddorwin@chromium.org>
Auto-Submit: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141097}