Use EXPECT_CALL properly in audio_context_manager_browsertest.cc
The documentation says "EXPECT_CALL must precede any code that
exercises the mock object." but this test has EXPECT_CALL in locations
after code that exercises the mock object.  Put the expectations
immediately after the mock object declaration instead.

This may be the source of flakiness in this test.  I'm not going to
re-enable the test in this CL, since we should wait until the flake
portal shows the flakiness is resolved.

Bug: 941219
Change-Id: I4a4858befa776e1b3f70f5d6209e9218a1125044
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4509910
Commit-Queue: Michael Wilson <mjwilson@chromium.org>
Reviewed-by: Sina Firoozabadi <sinafirooz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141111}