Added the mute state for audio sliders
Previous mute behavior: The icon on the slider is actually an ImageView.
It's not actionable. If the user clicks/taps the icon, it's the slider
that handles the event, and the slider goes to 0 and becomes muted.
Users can also drag the slider to the far left/press volume down key/
press mute key to mute the slider.

Current mute behavior: After this change, the icon becomes an
IconButton. If the user clicks/taps the icon, the slider becomes muted
while preserving the current level. Other mute behavior remains the
same.

1)Provided Style::kDefaultMuted and Style::kRadioActiveMuted for Quick
  Settings Slider.
2)Updated the main page volume slider to have the mute state.
3)Updated the audio subpage volume and mic sliders to have the mute
  state.
4)Changed the unit test for volume slider mute behavior.
5)Updated the font for the live caption label and slider labels.

Screenshots:
Volume slider in the main page:
https://screenshot.googleplex.com/AwD9XZsgAYduHhH
Interaction:
https://screenshot.googleplex.com/7Zvu6efZxR6oCgY

Sliders in the audio subpage:
https://screenshot.googleplex.com/8wpt9zM6Z7VLa6F
Interaction:
https://screenshot.googleplex.com/3QCWg6nznGyviWS

Live caption and slider label font update:
https://screenshot.googleplex.com/64vBPfHtdFmGYeK

Bug: b:274157670, b:278309598
Change-Id: Iaa7c7d0ed5f824d38d5a02d833bf7ca4ca12b382
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4475132
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Sylvie Liu <sylvieliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141112}