flags: improve error reporting for tests
The AboutFlagsTest that reports when flags are out of order does so by
telling the person running the test (the developer) that a flag is in
the wrong place. It should instead tell them what the right place for
the flag is, so that they can move it there.

This change:
* Modifies EnsureFlagsAreListedInAlphabeticalOrder() to report
  out-of-order flags as described above
* Removes the FlagFile type and uses of it from this test helper - it
  does not add any value over simply having constants to represent the
  filenames and passing those around
* Makes ReadFileContentsAsJSON() CHECK on invalid JSON instead of
  DCHECKing

Fixed: 1432086
Change-Id: I57248b70364a4100a1ce9c50ecc4f0e4e8e0f12d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4500338
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141138}