Prevent cross copy/move to ODFS when there is a managed user
Prevent files being copied or moved to ODFS if there is a managed user
present amongst other logged in users. This ensures that managed user's
files can't be leaked to a non-managed user's ODFS.

This situation can occur due to the "Multi-user sign-in" feature [1]
which is currently being deprecated. This change can be removed
(b/280947989) once the feature is fully deprecated.

[1] https://sites.google.com/corp/google.com/lacros/faqs/multi-something

Bug: b:278644796
Testing: Manual test
Change-Id: I46337eff132c27f5ac9df98da72a844653bf9117
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4505401
Reviewed-by: Jeremie Boulic <jboulic@chromium.org>
Commit-Queue: Cassy Chun-Crogan <cassycc@google.com>
Cr-Commit-Position: refs/heads/main@{#1141160}