Deflake and restructure navigation queueing tests.
The tests for navigation queueing were a little flaky under load for
subtle reasons: after beginning a renderer-initiated navigation, the
browser test would occasionally report that there was no speculative
RenderFrameHost.

It turns out that there was a subtle race: the browser test utils
provide a helper called `BeginNavigateToURLFromRenderer()`, which is
really just a wrapper around calling `ExecJs("window.location = ...")`.
What would happen during a race is:

1. The `ExecJS()` sends an IPC to execute the given JS snippet in the
   current RenderFrameHost, e.g. in the a.com renderer.
2. The a.com renderer calls the `BeginNavigation()` IPC and the browser
   process starts a navigation on its behalf.
3. The browser creates a speculative RenderFrameHost, e.g. a b.com
   renderer.

At this point, the a.com renderer has not sent a response to the initial
`ExecJS()` request, so the browser is still spinning an event loop
waiting for script execution in the a.com renderer to complete.

However, since the navigation will commit in the b.com renderer, the
navigation started in step 2 and step 3 above can complete *before* the
browser process ever hears about the result of the `ExecJS()` IPC from
the a.com renderer.

The fixes:

`BeginNavigateToURLFromRenderer()` no longer uses `ExecJS()`. Instead,
it uses `ExecuteScriptAsync()` and waits for a `DidStartNavigation()`
for that navigation before returning to the caller.

`BeginNavigationInCommitCallbackInterceptor` no longer exists: it was a
complex layer of nested waits that was quite difficult to understand,
and was very specific to the navigation queueing tests.  Instead, it has
been replaced by three smaller and more generic utilities that are also
more composable:

- `CommitNavigationPauser` is a scoper that waits for the renderer to
  send `DidCommitProvisionalLoad()` to the browser process, pauses it,
  and optionally, allows the paused commit to be resumed.
- `OnNextDidStartNavigation()` is a helper method that executes the
  given functor on the next `DidStartNavigation()` in a given
  `WebContents`.
- Finally, `ResumeCommitClosureSetWaiter` has been extracted from the
  original `BeginNavigationInCommitCallbackInterceptor` and is no longer
  responsible for resuming a paused commit , allowing the test logic to
  be a bit simpler and easier to read.

The smaller helper utilities also mean that the tests no longer have to
test RunLoops, which can lead to surprising failures and hangs.

Miscellaneous cleanups:

The tests themselves are now called "During" rather than "After" to more
accurately describe the problematic state. Renaming the tests hopefully
has the bonus side effect of having the CQ confirm that the "newly
added" tests are not flaky.

Note that there are still more opportunities for improving the browser
tests to match how such a race condition might play out in production:
one such opportunity would be to pause the IPC::ChannelProxy for a
RenderProcessHost when pausing the navigation commit process. This may
or may not be implemented in the future...

Bug: 1220337
Change-Id: If9752338eebef3a35354ec41df5348e78a7788f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4510362
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141190}