Fix flaky SiteInstanceTest.DoesSiteRequireDedicatedProcess test.
This test was flaky because it adds a custom WebUI scheme via
ContentBrowserClient::GetAdditionalWebUISchemes(), but the test did
not actually see the effect of this due to the caching done by
URLDataManagerBackend::GetWebUISchemes().  So, if some other unit test
ran in the same shard as this test and triggered the caching to happen
sooner (prior to the custom WebUI scheme being added), this test was
seeing the stale list of WebUI schemes when checking it in
SiteInfo::RequiresDedicatedProcess().  This only mattered on Android
since SiteInfo::RequiresDedicatedProcess() returned true sooner on
desktop (from UseDedicatedProcessesForAllSites()).

Ideally, all unit tests should run with this kind of caching
disabled. For now, though, given that there's only one test that
overrides GetAdditionalWebUISchemes(), add some plumbing so that it
disables the caching.

Bug: 1377466
Change-Id: I116ace22c7c8edd470d4b57d20ba7207634237d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4514537
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141200}