[FilesBulkPinning] Show tick for folders when bulk pinning syncing
When bulk pinning is enabled all items are downloaded, however, folders
are excluded. This is to ensure items that are updated in the cloud can
first be verified that they can be fit on the disk before actually
pinning them. This leaves folders without the offline tick, even though
their contents are all pinned offline and new contents are made
available offline (same behaviour as pinning).

In this way, this ensures the tick appears on all folders when the bulk
pinning manager is in the syncing state. The existing behaviour should
be retained if the user disables folder syncing (i.e. if a folder is
explicitly pinned it should retain the folder tick, otherwise it should
not have one). Furthermore, shortcuts should are excluded from being
pinned and traversed so they do not receive any special treatment when
the pin manager is enabled.

Bug: b:278002577
Test: browser_tests --gtest_filter=*drive*
Change-Id: I348fb89db7c1840c109915e9445f11b8255e39e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4505998
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141206}