Prerender: Store prerendering requests from background pages when the trigger is speculation rules
Before this CL, any prerendering requests were canceled and never had a
chance to be prerendered if the trigger was in the background at the
time `CreateAndStartHost()` was called.

This CL allows to store prerendering requests(represented as
`PrerenderHost`) from background pages in the `pending_prerenders` queue
and pop/handle them when the pages become foreground on
`StartPrerendering()`.

Since the current implementation only uses `pending_prerenders` queue
when the trigger type is speculation rules, this CL only targets this.
No changes are intended when the trigger is embedder.


Bug: 1433602
Change-Id: I1b0d107fde16d69b0c7828cb3e9d7a904fd076ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4489084
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Huanpo Lin <robertlin@chromium.org>
Commit-Queue: Taiyo Mizuhashi <taiyo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141277}