Fix flaky PasswordProtection test
This de-flakes the DeferActivatePrerenderStartedAfterRequest test. This
test would occasionally fail as the PasswordProtectionRequest issues a
real request using the SimpleURLLoader. These browser tests expect that
a request won't receive a response (in time?) and mock responses by
manually calling FinishRequest. However, in the flaky runs the URLLoader
receives a real response before the test's mock response and this causes
the CommitDeferringCondition to be destroyed unexpectedly.

This CL adds a bit to PasswordProtectionRequest that allows skipping the
URLLoader request from tests.

Bug: 1433599
Change-Id: Id57f308f114043e321f54a57f9854145dbd56ec5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4509446
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141309}