[iOS] Fix the Password Bottom Sheet’s title background color
In dark mode, the background color of the title view was
different from the background color of the rest of the
bottom sheet. This was not the case without dark mode.
The title view’s background color was adjusted, along
with its parent view’s background color.

Bug: 1442504
Change-Id: Ibf473aca622f5be7168344b167e569bcdd96639e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4505161
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Alexis Hétu <sugoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141311}