Fix a bug that swapped GenerateBid dependency latency metics.
A bug in BidderWorklet::DeliverBidCallbackOnUserThread inadvertently
swapped a pair of the GenerateBidDependencyLatencies. There was test
coverage, but the tests were unfortunately also wrong. As a result,
promises latency was recorded as trusted_signals latency and
trusted_signals latency was recorded as promises latency. This affected
the following metrics:
- MeanGenerateBidTrustedBiddingSignalsLatencyInMillis
- MaxGenerateBidTrustedBiddingSignalsLatencyInMillis
- MeanGenerateBidTrustedBiddingSignalsCriticalPathLatencyInMillis
- NumGenerateBidTrustedBiddingSignalsOnCriticalPath
- MeanGenerateBidConfigPromisesLatencyInMillis
- MaxGenerateBidConfigPromisesLatencyInMillis
- MeanGenerateBidConfigPromisesCriticalPathLatencyInMillis
- NumGenerateBidConfigPromisesOnCriticalPath

To account for the changed meaning of some of these metrics, this change
renames the `AdsInterestGroup.AuctionLatency` metric to `AdsInterestGroup.AuctionLatency.V2`. Any analysis that doesn't use the above metrics can UNION ALL these events together to analyze more data during the rollout of this fix.

Change-Id: Ibe4d98e76dc71e76f343156ce86185f5599d30dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4509621
Reviewed-by: Sun Yueru <yrsun@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Orr Bernstein <orrb@google.com>
Cr-Commit-Position: refs/heads/main@{#1141327}