Remove NetworkDelegate::IsShowingNetworkConfigScreen
In a previous CL, we ensure that KioskAppLauncher can never be running
whilst the network screen is showing, so we no longer need to check for
this in KioskAppLauncher or its callbacks to KioskLaunchController.

Additionally, we do not need to check this in tests, as we're already
waiting for the presence of the network screen.

Bug: b:256596278
Test: KioskLaunchControllerTest
Test: KioskDeviceOwnedTest
Test: KioskEnterpriseTest
Change-Id: I2d7361e40d4a0005234d2539b9d944ed36189319
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4507256
Reviewed-by: Jeroen Dhollander <jeroendh@google.com>
Commit-Queue: Ben Franz <bfranz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141356}