[ios] Fix TwoScreensSigninCoordinator to handle interruption
When the coordinator is interrupted, the child coordinator needs to be
stopped and cleaned up. When the user takes an action on the UI to
end the signin operation, the child coordinator is cleaned up earlier.
This change allows the coordinator to handle either scenario.

Bug: 1427439
Change-Id: I93cc3c84415c3cc844e8b00a47f85c85d7c03661
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4509978
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1141392}