Tweak net-internals HTTPS-RR support
Noticed this in the process of aligning these with the spec.
net_internals_ui.cc was serializing some information that the viewer
wasn't doing anything with. But, in serializing the ECHConfigList (a
byte string), it turned each byte into a separate JS integer, which
is a bit unusual. Instead, just base64-encode it, which should be
more compact.

Next, rename, at least in the net-internals code, "endpoints with
metadata" to "alternative endpoints", which matches the spec's
terminology. See https://crbug.com/1416410#c8. A follow-up CL will do a
mass rename across the codebase. I've separated these out because there
is an actual behavior change.

This CL also tweaks the output in chrome://net-internals. The old output
was focused on ALPN, but HTTPS-RR/SVCB does a lot more than that. It
also uses a somewhat awkward phrasing of "supported protocol alpns".
Instead, simplify the JSON output slightly and just dump the JSON
output. We seem to already be happy outputting raw JSON, so may as well
do that throughout. Introduce each entry as an "alternative endpoint"
per the spec's terminology.

Finally, use Array.forEach instead of Array.map. When not making another
array, forEach is the method to use.

I've also trimmed the names of the tests slightly. The references to
"single host" and "multiple host" don't seem to convey much. (There's
just one host. The "multiple host" seems to refer to there being
multiple IPs.)

Bug: 1416410, 1345522
Change-Id: I4cf48b28e9f9e677f7ef3cb2ceda70e4c2e6810f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4505348
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141437}