PrivacyIndicators: Adjust NumberOfRepeatedShows metric
Record proper shows so we can analyze the repeated shows in context.

Also, we record more data for this metric so that the histograms can be
generated properly. This metric doesn't have a lot of data at the moment
since it is for recording bad signal only.

Change-Id: I06f65eccf213da407b5000466f4fe77960c88cd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4515559
Commit-Queue: Andre Le <leandre@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141439}