Fix a crash in SubframeHistoryNavigationThrottle
This happens when multiple frames traverse simultaneously, and the main
frame has a navigate event handler that might cancel the navigation.
If an iframe is navigating cross-document and has a slow beforeunload
event handler, it may not attempt to create throttles until after the
main frame traversal has completed. In that case, it will try to get the
main frame NavigationRequest that no longer exists and crash.

Fixed: 1443312
Change-Id: I79b26361128026bf13cbe86c26d4f33749c043d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4515156
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141473}