[iOS] Remove unused code paths in startup.
This CL follows-up on crbug.com/1346512, which added assertions to verify that some code paths in the startup sequence were never used.
As far as we can tell after ~10 months, those code paths are in fact
never used, so this CL removes them along with a surprising amount of
other attached logic.

Specifically:

In [AppState applicationWillEnterForeground:], there is code that
handles the case where the 'Start' init stage hasn't happened yet. This
is now just an assertion and the code is removed.

All of the code that handles the 'launchOptions' from
[UIApplicationDelegate application:didFinishLaunchingWithOptions:]
is removed. The launch options aren't being read by any code, so
there's no need to save them. There were conditions around whether
the app was in the background when they were saved, but the app is
always in the background at this point. The only thing that needs
to happen at the point that these options were handled is that the
app state needs to start its init cycle; that's now an explicit
method.

Accordingly, the LaunchOptions property in the BrowserLauncher and
StartupInformation protocols (which had the *same implementation!*)
is removed, as is its definition. There were some big hairy tests
to verify that AppState was doing the right thing in this process, and
those are also now gone.

Bug: 1346512
Change-Id: I6c6d1afcc830bb6f200a290959945af00b43147f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4499106
Reviewed-by: Federica Germinario <fedegermi@google.com>
Reviewed-by: edchin <edchin@google.com>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141482}