Avoid multiple tab count writes to SharedPreferences in MultiInstanceManager during app startup
MultiInstanceManagerApi31#initialize() adds a
TabModelSelectorTabModelObserver during app startup in
ChromeTabbedActivity#createTabModels() that causes multiple writes of
the tab count to SharedPreferences for every tab that is restored on
start. This CL defers the addition of this observer to after the tab
state is initialized to optimize this so that the tab count is written
once after the tab state is initialized.

Bug: 1441767
Change-Id: I0684504eb7dd517bce9c1b560df6727dfcd06f1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4515161
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Aishwarya Rajesh <aishwaryarj@google.com>
Cr-Commit-Position: refs/heads/main@{#1141542}