Refactor SubprocessMetricsProvider
This CL is gated behind a feature flag, and logically, it should be a
no-op (given that the feature is disabled).

When the feature is enabled, SubprocessMetricsProvider is leaky. It
exists throughout the whole lifetime of Chrome, and is always there to
listen for subprocesses exiting (to merge their deltas one last time).
When the feature is disabled, SubprocessMetricsProvider is owned by
MetricsService, and so if a subprocess exits after MetricsService was
destroyed, the exit notification is "lost" and the last metrics are
not merged.

On my Linux dev machine, I can confirm that enabling the feature
(making SubprocessMetricsProvider leaky) makes it so that we now
receive all notifications before Chrome terminates. Note that because
these subprocesses exit after MetricsService is destroyed, the metrics
do not make it into the last log. Instead, they will be recovered by
the persistent histogram system in some followup session. Since
Android does not make use of a clean shutdown path, I do not think
that this feature makes any difference at all on that platform.

---

Other refactors:
 - Moved MergeHistogramDeltasFromAllocator() to anon namespace, since
   the end goal is calling this on a background thread.
 - Changed the internal map that maps a process ID to its allocator
   from base::IDMap to std::map. The former uses std::unordered_map,
   which is usually not recommended. Further, changing to std::map
   allows for fewer lookups in the code.
 - Removed RegisterExistingRenderProcesses(). Instead,
   SubprocessMetricsProvider is instantiated earlier, such that we
   ensure that no child process exist at that time.

---

Here's a "one-pager" doc with more details:
https://docs.google.com/document/d/1pIKCfYvopOaBR-6XgRMtGPQhzh17_tggT77NsjfThyU/edit?usp=sharing&resourcekey=0-abaTO_zMImjbqyHxWVM7_w

Bug: 1293026
Change-Id: I1169a20ddfa89eef6ad2be1c0b85700e08a33938
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4490523
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141555}