[Tab Switcher] Remove global layout listener
As far as I can determine this was only used for caching the rect of
the thumbnail for the tab shrink animation. Now that this animation is
deferred until layout is completed we can just update the value
dynamically rather than caching it.

This CL also prevents layout updates to the TabListContainerView while
it is invisible. This avoid visual janks during the transition
animation. This also avoid extra layout calls whenever the top controls
offset is changed which should hopefully help a bit with scroll jank.

Bug: 1327694, 1443639
Change-Id: Ifb46e16c9d2e9a370b580bc11b8e756e51407e85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4496674
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Mei Liang <meiliang@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1141575}