[Shopping] Add experiment for recording PDP state and detection method
This patch adds an experiment to start recording potential PDPs that
are detected locally. This requires running the metadata extraction
javascript on every page rather than just on server-detected PDPs. The
page is currently detected to be a PDP locally if it has one or more
of various open graph meta tags in the header -- this includes price,
currency, type, and image availability.

This metric is recorded for both users with and without the
experiment. The difference is that users without the experiment
will not run the on-page extraction to run on every page. This
means we'll only see reports for "server only" and "both server
and local".

Bug: b:278291710
Change-Id: Iae8cda6177abbf934c41f66e263f0dd1ffe61654
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4426168
Reviewed-by: Ayman Almadhoun <ayman@chromium.org>
Reviewed-by: Yue Zhang <yuezhanggg@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141592}