Fix assistant shadow animation being misplaced behind container view
There is inaccuracy when animating a shadow layer with both a scale and
transform at the same time. To fix this, inset the `from_rect` so that
the transformed shadow is completely hidden behind the view layer at
the start of animation and slowly reveals itself when animating to the
proper size.

Bug=b:278531137

Change-Id: I8692bb5a8e2a936af6347cae984d877864cf90a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4492148
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141597}